requirements.txt
fastapi==0.109.2
uvicorn~=0.30.1
requests~=2.32.3
jinja2==3.1.2
python-dotenv~=1.0.1